requirements.txt
pyserial==3.5
pyzmq==22.2.1
inputs==0.5
